requirements.txt
Django>=4.1.3
djangorestframework==3.14.0
django-cors-headers==4.2.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-dotenv==1.0.0
django-filter==23.2
django-geojson==4.0.0
djangorestframework-gis==1.0
